Type linter, removed a redundant function
